Update LICENSE.md
I don't believe it is good practice to modify the default MIT license (e.g. put it in quote blocks) unless you are a copyright lawyer. One practical downside of this behavior is that Github does not identify it as the MIT license and therefore does not display the useful summary banner on top. This practice is widespread in the Julia ecosystem, and since a lot of people model their package after this one, I thought it might be worth fixing it here. Please ignore if you disagree. Thanks!